backend/requirements.txt
numpy<2
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
gitpython==3.1.43
groq==0.9.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
aiofiles==23.2.1
python-dotenv==1.0.1
tiktoken==0.7.0
pygments==2.18.0
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.2
pytest-asyncio==0.23.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
smolagents==1.1.0
mcp==1.2.1
